requirements.txt
absl-py==2.1.0
accelerate==0.24.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
autopep8==2.0.4
backgroundremover==0.2.6
beautifulsoup4==4.12.2
blinker==1.8.2
body-measurements==1.1.1
bottleneck==1.3.7
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chess==1.10.0
chumpy==0.70
clang==5.0
clarabel==0.6.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.2.0
commandlines==0.4.1
contourpy==1.2.0
cvxpy==1.5.0
cvzone==1.6.1
cycler==0.10.0
cython==3.0.5
debugpy==1.8.0
decorator==4.4.2
detectors==0.1.10
dnspython==2.4.2
easydict==1.11
easygui==0.98.3
ecos==2.0.12
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.104.1
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
filelock==3.9.0
filetype==1.2.0
flask==3.0.3
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
future==0.18.3
fuzzy-c-means==1.7.0
fvcore==0.1.5.post20221221
gast==0.4.0
gcloud==0.18.3
gdown==4.7.1
gensim==4.3.2
google-auth-oauthlib==1.0.0
google-auth==2.23.4
google-pasta==0.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.1
grpcio==1.59.2
h11==0.14.0
h5py==3.11.0
hsh==1.1.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.18.0
humanfriendly==10.0
idna==2.10
imageio-ffmpeg==0.4.9
imageio==2.32.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
imutils==0.5.4
iopath==0.1.10
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
jsonschema-specifications==2023.7.1
jsonschema==4.19.2
jupyter-client==8.6.0
jupyter-core==5.5.0
jws==0.1.3
keras-applications==1.0.8
keras-preprocessing==1.1.2
keras==3.3.3
kiwisolver==1.4.5
lazy-loader==0.3
libclang==16.0.6
littleutils==0.2.2
llvmlite==0.41.1
lxml==4.9.3
mako==1.2.4
markdown-it-py==3.0.0
markdown==3.5.1
markupsafe==2.1.3
matplotlib-inline==0.1.6
matplotlib==3.8.1
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
mock==5.1.0
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
mysql-connector-python==9.0.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.5.8
networkx==3.0
nltk==3.8.1
numba==0.58.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
ogb==1.3.6
onnxruntime==1.16.3
opencv-contrib-python==4.8.1.78
opencv-python-headless==4.8.0.74
opencv-python==4.9.0.80
opendr==0.73
openpyxl==3.1.3
opt-einsum==3.3.0
optree==0.11.0
optuna==3.4.0
orjson==3.9.10
osqp==0.6.3
outdated==0.2.2
packaging==23.2
pafy==0.5.5
pandas==2.1.2
parso==0.8.3
pillow==9.3.0
pip==24.0
platformdirs==3.11.0
plotly==5.21.0
pooch==1.8.0
portalocker==2.8.2
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1-modules==0.3.0
pyasn1==0.5.0
pybind11==2.11.1
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pydantic-core==2.14.6
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic==2.5.3
pygments==2.16.1
pymatting==1.1.12
pyopengl-accelerate==3.1.6
pyopengl==3.1.6
pyparsing==3.1.1
pyproject-toml==0.0.10
pyrealsense2==2.54.1.6182
pyrebase==3.0.27
pysmartdl==1.3.1
pysocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.1.1
pytorch-metric-learning==2.3.3
pytorch3d==0.7.0
pytorchcv==0.0.67
pytz==2023.3.post1
pyunpack==0.3
pywavelets==1.4.1
pyyaml==6.0.1
regex==2023.10.3
reportlab==4.0.7
requests-oauthlib==1.3.1
requests==2.31.0
rich==13.5.2
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.3
sentencepiece==0.1.99
setproctitle==1.3.2
setuptools==68.1.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.32.0
sympy==1.12
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboard==2.14.1
tensorboardx==2.6.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-gpu==2.11.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-io==0.35.0
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2023.8.12
timm==0.9.5
tokenizers==0.14.1
tomli==2.0.1
torch==2.1.0
torchsummary==1.5.1
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
triton==2.1.0
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0
virtualenv==20.24.5
websocket-client==1.7.3
websockets==11.0.3
werkzeug==3.0.1
win32-setctime==1.1.0
win32-setfileattributes==0.2.0
wordcloud==1.8.2.2
xgboost==2.1.0
yapf==0.40.2
yarl==1.9.3
zipp==3.17.0
roboflow